requirements.txt
mysql-connector-python==8.0.28
protobuf==3.19.4
python-dotenv==0.20.0
